chore: remove secp256k1 wasm compatable
