+++ secio/Cargo.toml
@@ -28,6 +28,7 @@ molecule = { version = "0.6.0", optional = true }
 
 unsigned-varint = "0.3"
 bs58 = "0.3.0"
+secp256k1 = "0.19"
 
 [target.'cfg(unix)'.dependencies]
 openssl = "0.10.25"
@@ -35,12 +36,10 @@ openssl-sys = "0.9"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 rand = "0.7"
-secp256k1 = "0.17"
 ring = "0.16.5"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 rand = { version = "0.7", features = ["wasm-bindgen"] }
-secpwasm = { version = "0.3.5", package = "libsecp256k1"}
 sha2 = "0.9.0"
 hmac = "0.9.0"
 x25519-dalek = "1.1"
@@ -50,7 +49,6 @@ chacha20poly1305 = "0.7"
 env_logger = "0.6"
 criterion = "0.3"
 tokio = { version = "0.2.0", features = ["tcp", "rt-core", "dns"] }
-secpwasm = { version = "0.3.5", package = "libsecp256k1"}
 sha2 = "0.9.0"
 hmac = "0.9.0"
 x25519-dalek = "1.1"
